refactor: replace MessageRepository partial classes with 3 distinct classes and interfaces
Split IMessageRepository into 3 focused interfaces:
- IMessageRepository: core CRUD (AddAsync, GetByIdAsync, UpdateAsync, SoftDelete, etc.)
- IMessagePaginationRepository: GetChannelPageAsync, GetConversationPageAsync
- IMessageSearchRepository: SearchGuildMessagesAsync, SearchConversationMessagesAsync

Each backed by its own implementation class in Persistence/Messages/.
Shared logic (attachment queries, mapping, reaction dict building) extracted
into internal static MessageRepositoryHelpers — no duplication.

Inner row types (MessageAttachmentRow, ReactionSummaryRow) promoted to
internal classes in Rows/ alongside other row types.

Handlers and unit test mocks updated to inject the minimal interface needed.

Closes #244
