+++ src/Harmonie.Infrastructure/Persistence/Messages/MessageRepositoryHelpers.cs
@@ -1,0 +1,114 @@
+using Dapper;
+using Harmonie.Application.Interfaces;
+using Harmonie.Domain.Entities;
+using Harmonie.Domain.ValueObjects;
+using Harmonie.Infrastructure.Rows;
+
+namespace Harmonie.Infrastructure.Persistence.Messages;
+
+internal static class MessageRepositoryHelpers
+{
+    internal static async Task<IReadOnlyDictionary<Guid, IReadOnlyList<MessageAttachment>>> GetAttachmentsByMessageIdsAsync(
+        DbSession dbSession,
+        IReadOnlyCollection<Guid> messageIds,
+        CancellationToken cancellationToken)
+    {
+        if (messageIds.Count == 0)
+            return new Dictionary<Guid, IReadOnlyList<MessageAttachment>>();
+
+        const string sql = """
+                           SELECT ma.message_id AS "MessageId",
+                                  ma.position AS "Position",
+                                  uf.id AS "UploadedFileId",
+                                  uf.filename AS "FileName",
+                                  uf.content_type AS "ContentType",
+                                  uf.size_bytes AS "SizeBytes"
+                           FROM message_attachments ma
+                           INNER JOIN uploaded_files uf ON uf.id = ma.uploaded_file_id
+                           WHERE ma.message_id = ANY(@MessageIds)
+                           ORDER BY ma.message_id, ma.position
+                           """;
+
+        var connection = await dbSession.GetOpenConnectionAsync(cancellationToken);
+        var command = new CommandDefinition(
+            sql,
+            new { MessageIds = messageIds.ToArray() },
+            transaction: dbSession.Transaction,
+            cancellationToken: cancellationToken);
+
+        var rows = await connection.QueryAsync<MessageAttachmentRow>(command);
+
+        return rows
+            .GroupBy(row => row.MessageId)
+            .ToDictionary(
+                group => group.Key,
+                group => (IReadOnlyList<MessageAttachment>)group
+                    .OrderBy(row => row.Position)
+                    .Select(row => new MessageAttachment(
+                        UploadedFileId.From(row.UploadedFileId),
+                        row.FileName,
+                        row.ContentType,
+                        row.SizeBytes))
+                    .ToArray());
+    }
+
+    internal static IReadOnlyDictionary<Guid, IReadOnlyList<MessageAttachment>> BuildAttachmentsDictionary(
+        IEnumerable<MessageAttachmentRow> rows)
+    {
+        return rows
+            .GroupBy(row => row.MessageId)
+            .ToDictionary(
+                group => group.Key,
+                group => (IReadOnlyList<MessageAttachment>)group
+                    .OrderBy(row => row.Position)
+                    .Select(row => new MessageAttachment(
+                        UploadedFileId.From(row.UploadedFileId),
+                        row.FileName,
+                        row.ContentType,
+                        row.SizeBytes))
+                    .ToArray());
+    }
+
+    internal static IReadOnlyDictionary<Guid, IReadOnlyList<MessageReactionSummary>> BuildReactionsDictionary(
+        IEnumerable<ReactionSummaryRow> rows)
+    {
+        return rows
+            .GroupBy(row => row.MessageId)
+            .ToDictionary(
+                group => group.Key,
+                group => (IReadOnlyList<MessageReactionSummary>)group
+                    .Select(row => new MessageReactionSummary(
+                        row.Emoji,
+                        row.Count,
+                        row.ReactedByCaller))
+                    .ToArray());
+    }
+
+    internal static Message MapToMessage(
+        MessageRow row,
+        IReadOnlyDictionary<Guid, IReadOnlyList<MessageAttachment>> attachmentsByMessageId)
+    {
+        var contentResult = MessageContent.Create(row.Content);
+        if (contentResult.IsFailure || contentResult.Value is null)
+            throw new InvalidOperationException("Stored message content is invalid.");
+
+        GuildChannelId? channelId = row.ChannelId.HasValue
+            ? GuildChannelId.From(row.ChannelId.Value)
+            : null;
+        ConversationId? conversationId = row.ConversationId.HasValue
+            ? ConversationId.From(row.ConversationId.Value)
+            : null;
+        attachmentsByMessageId.TryGetValue(row.Id, out var attachments);
+
+        return Message.Rehydrate(
+            MessageId.From(row.Id),
+            channelId,
+            conversationId,
+            UserId.From(row.AuthorUserId),
+            contentResult.Value,
+            row.CreatedAtUtc,
+            row.UpdatedAtUtc,
+            row.DeletedAtUtc,
+            attachments);
+    }
+}
